First working version of the new overview UI.
Moved OverviewActivity to EmpaticaE4 package and transferred
MainActivity functions to OverviewActivity.
